// packages/agno-api/src/generated/routes.gen.ts
export type ContentType = 'application/json' | 'multipart/form-data' | 'application/x-www-form-urlencoded'

export interface RouteMeta {
  readonly query: readonly string[]
  readonly body: readonly string[]
  readonly contentType: ContentType | null
}